fix borg ui mispredict opening (#28305)
move borg ui junk to shared
