+++ Content.Shared/Wires/ActivatableUIRequiresPanelComponent.cs
@@ -1,15 +1,17 @@
-﻿namespace Content.Server.Wires;
+﻿using Robust.Shared.GameStates;
+
+namespace Content.Shared.Wires;
 
 /// <summary>
 /// This is used for activatable UIs that require the entity to have a panel in a certain state.
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent, Access(typeof(SharedWiresSystem))]
 public sealed partial class ActivatableUIRequiresPanelComponent : Component
 {
     /// <summary>
     /// TRUE: the panel must be open to access the UI.
     /// FALSE: the panel must be closed to access the UI.
     /// </summary>
-    [DataField("requireOpen"), ViewVariables(VVAccess.ReadWrite)]
+    [DataField]
     public bool RequireOpen = true;
 }
